Refactor restart property checking and only handle restart checking when viamserver does not
